feat(engine): Account abstraction, hold lifecycle, and the spine account-projection drive (ADR-PC-033 / ADR-PC-032 §A1)
In plain English: the engine can now treat any product instance as an account
whose balances are always recomputed from recorded facts, never stored as a
mutable number. Approving a card-style debit sets money aside as a "hold" that
lowers what is spendable until the payment settles or the hold times out; and a
new background drive finally feeds the movement ledger (and the new hold set)
from real appended events, so account statements and balances fill up at
runtime instead of only in unit tests.

The engine half of the ADR-PC-033 Account contract, plus the ADR-PC-032 §A1
live projection drive it folds through:

- IAccount / IHoldable family-declared seams + the spine Hold value object
  (engine/src/Babelstone.Engine/Account.cs): the IErasable-style role
  declaration — the engine owns the generic fold, a family supplies only its
  state and account_ref. AccountBalanceReader computes
  accounting = Σ(signed movements) and available = accounting − Σ(active
  holds); both reads, never columns (ACCOUNT_BALANCE_IS_A_FOLD).
- operations.HoldPlaced / HoldCaptured / HoldExpired cross-cutting event
  records + generic no-op per-family folds (AccountHoldEvents.cs), spliced
  into CrossCuttingEventRegistrations so every family stream decodes them.
  Store-only per ADR-IC-017 (no named consumer yet); their governed Avro wire
  shapes are authored dormant as contracts/avro/operations/Hold*.avsc.json
  (the _shared/Movement.avsc.json posture) so bus promotion is a rename, not
  a reshape.
- AccountHoldProjector + IAccountHoldStore/PostgresAccountHoldStore +
  migration 0020 (account_holds): the rebuildable active-hold fold, idempotent
  on hold_id (a duplicate release transitions zero rows — no double-release),
  with the ADR-PC-023 projection-derived expiry-horizon read (the horizon is
  an input, never a clock).
- FundsAndRulesDecider (AccountAuthorization.cs): the pure ADR-PC-030 §P3
  stages 3–5 — funds vs available balance, pack rules (per-transaction limit,
  descoberto autorizado overdraft), then the HoldPlaced earmark. Declines
  carry a generic reason taxonomy; the decider never appends on refusal.
- SpineProjectionDrainer + ISpineProjector + SpineProjectionRelayService:
  the production caller for MovementLedgerProjector.ApplyAsync (and the hold
  projector) — a spine singleton at the composition root, NOT the per-family
  ProjectionDrainer, because the read models are account-keyed and
  cross-family. Per-stream tail drain with project-then-checkpoint under the
  shared spine.account_ledger kind; decode goes through each family's own
  IFamilyModule bindings (a merged registry would collide on the shared
  operations.* types).
- Fix: PostgresMovementLedgerStore.GetBalanceCentsAsync always returned 0
  against real PostgreSQL — SUM(bigint) is NUMERIC, so the scalar never
  pattern-matched long. Caught by the new integration suite; fixed with a
  ::bigint cast (same guard applied to the new hold-cents sum).
- Tests: AccountHoldProjectorTests (HOLD_LIFECYCLE_PURE: three transitions,
  hold_id idempotency, partial capture, replay determinism),
  FundsAndRulesDeciderTests, and SpineAccountProjectionIntegrationTests
  (Testcontainers): real appends → ledger statement + signed-sum balance,
  hold lifecycle → available balance, at-least-once redelivery no-op, and a
  byte-identical truncate+refold rebuild. account_holds added to the
  EventStoreSchemaFamilyAgnosticTests allowlist (a deliberate generic-engine
  edit).

Honours ADR-PC-033, ADR-PC-032 (§A1 seam + §A5 idempotency), ADR-PC-030 §P3,
ADR-PC-034 (sync technique), ADR-PC-023 (no clock-manufactured expiry),
ADR-PC-021 §P2 (the seams, events, and drive name no family), ADR-IC-017
(store-only until a named consumer), ADR-PC-010 (pure folds, integer cents).

Closes-on-merge: babelstone-hkwp.3, babelstone-hkwp.5 (recorded in the PR).

+++ engine/src/Babelstone.EventStore/MovementLedger.cs
@@ -140,9 +140,11 @@ public sealed class PostgresMovementLedgerStore(string connectionString) : IMove
     public async Task<long> GetBalanceCentsAsync(string accountRef, CancellationToken ct = default)
     {
         // The balance is the signed sum: Credit adds, Debit subtracts (direction is relative to the
-        // account). COALESCE makes an account with no rows read as zero, not NULL.
+        // account). COALESCE makes an account with no rows read as zero, not NULL. The ::bigint cast
+        // is load-bearing: PostgreSQL's SUM(bigint) returns NUMERIC, which Npgsql surfaces as
+        // decimal — without the cast the scalar read below would never see an Int64.
         const string sql = """
-            SELECT COALESCE(SUM(CASE WHEN direction = 'Credit' THEN amount_cents ELSE -amount_cents END), 0)
+            SELECT COALESCE(SUM(CASE WHEN direction = 'Credit' THEN amount_cents ELSE -amount_cents END), 0)::bigint
             FROM movement_ledger
             WHERE account_ref = @account_ref;
             """;
